+++ core/src/com/giancola/battleship/Util.kt
@@ -6,3 +6,5 @@ import kotlin.random.Random
 val rnd = Random.Default
 
 fun randomColor(): Color = Color(rnd.nextFloat(), rnd.nextFloat(), rnd.nextFloat(), 1.0f)
+
+fun coords2str(gridX: Int, gridY: Int): String = "${(gridX + 'A'.toInt()).toChar()}${gridY+1}"